refactor(db/seed): move user data to model file
